배포 전 완성본

+++ src/main/java/com/study/backendeum/entity/User_Entity.java
@@ -1,5 +1,6 @@
 package com.study.backendeum.entity;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
@@ -12,7 +13,7 @@ import lombok.Data;
 public class User_Entity {
     @Id
     @Column(name = "student_id")
-    private int student_id;
+    private Long student_id;
 
     @Column(length = 20, nullable = false)
     private String student_name;
