+++ src/main/java/com/study/backendeum/repository/User_Repository.java
@@ -3,6 +3,6 @@ package com.study.backendeum.repository;
 import com.study.backendeum.entity.User_Entity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface User_Repository extends JpaRepository<User_Entity, Integer> {
+public interface User_Repository extends JpaRepository<User_Entity, Long> {
 
 }
